refactor: update code structure
