+++ packages/relay/src/lib/services/ethService/ethCommonService/ICommonService.ts
@@ -18,5 +18,17 @@
  *
  */
 
-export * from './ethCommonService';
-export * from './ethFilterService';
+import {JsonRpcError} from "../../../errors/JsonRpcError";
+
+export interface ICommonService {
+
+  blockTagIsLatestOrPending(tag: any): boolean;
+
+  validateBlockRangeAndAddTimestampToParams(params: any, fromBlock: string, toBlock: string, requestIdPrefix?: string): Promise<boolean>;
+
+  getHistoricalBlockResponse(blockNumberOrTag?: string | null, returnLatest?: boolean, requestIdPrefix?: string | undefined): Promise<any | null>;
+
+  getLatestBlockNumber(requestIdPrefix?: string): Promise<string>;
+
+  genericErrorHandler(error: any, logMessage?: string): JsonRpcError;
+}
