+++ packages/relay/src/lib/services/ethService/ethFilterService/IFilterService.ts
@@ -18,6 +18,10 @@
  *
  */
 
+import { JsonRpcError } from "../../../errors/JsonRpcError";
+
 export interface IFilterService {
+  newFilter(fromBlock: string, toBlock: string, address?: string, topics?: any[], requestIdPrefix?: string):  Promise<string | JsonRpcError>;
+
   uninstallFilter(filterId: string, requestId?: string): Promise<boolean>;
 }
