+++ packages/server/src/server.ts
@@ -168,7 +168,7 @@ app.getKoaApp().use(async (ctx, next) => {
   }
 
   let id = '';
-  
+
   if (options.query) {
     id = options.query as string;
   }
@@ -264,7 +264,7 @@ app.useRpc('eth_blockNumber', async () => {
  */
 app.useRpc('eth_estimateGas', async (params: any) => {
   return logAndHandleResponse('eth_estimateGas', params, (requestId) =>
-    relay.eth().estimateGas(params?.[0], params?.[1], requestId));
+      relay.eth().estimateGas(params?.[0], params?.[1], requestId));
 });
 
 /**
@@ -276,7 +276,7 @@ app.useRpc('eth_estimateGas', async (params: any) => {
  */
 app.useRpc('eth_getBalance', async (params: any) => {
   return logAndHandleResponse('eth_getBalance', params, (requestId) =>
-    relay.eth().getBalance(params?.[0], params?.[1], requestId));
+      relay.eth().getBalance(params?.[0], params?.[1], requestId));
 });
 
 /**
@@ -288,7 +288,7 @@ app.useRpc('eth_getBalance', async (params: any) => {
  */
 app.useRpc('eth_getCode', async (params: any) => {
   return logAndHandleResponse('eth_getCode', params, (requestId) =>
-    relay.eth().getCode(params?.[0], params?.[1], requestId));
+      relay.eth().getCode(params?.[0], params?.[1], requestId));
 });
 
 /**
@@ -298,7 +298,7 @@ app.useRpc('eth_getCode', async (params: any) => {
  */
 app.useRpc('eth_chainId', async () => {
   return logAndHandleResponse('eth_chainId', [], (requestId) =>
-    relay.eth().chainId(requestId));
+      relay.eth().chainId(requestId));
 });
 
 /**
@@ -310,7 +310,7 @@ app.useRpc('eth_chainId', async () => {
  */
 app.useRpc('eth_getBlockByNumber', async (params: any) => {
   return logAndHandleResponse('eth_getBlockByNumber', params, (requestId) =>
-    relay.eth().getBlockByNumber(params?.[0], Boolean(params?.[1]), requestId));
+      relay.eth().getBlockByNumber(params?.[0], Boolean(params?.[1]), requestId));
 });
 
 /**
@@ -322,7 +322,7 @@ app.useRpc('eth_getBlockByNumber', async (params: any) => {
  */
 app.useRpc('eth_getBlockByHash', async (params: any) => {
   return logAndHandleResponse('eth_getBlockByHash', params, (requestId) =>
-    relay.eth().getBlockByHash(params?.[0], Boolean(params?.[1]), requestId));
+      relay.eth().getBlockByHash(params?.[0], Boolean(params?.[1]), requestId));
 });
 
 /**
@@ -343,7 +343,7 @@ app.useRpc('eth_gasPrice', async () => {
  */
 app.useRpc('eth_getTransactionCount', async (params: any) => {
   return logAndHandleResponse('eth_getTransactionCount', params, (requestId) =>
-    relay.eth().getTransactionCount(params?.[0], params?.[1], requestId));
+      relay.eth().getTransactionCount(params?.[0], params?.[1], requestId));
 });
 
 /**
@@ -354,7 +354,7 @@ app.useRpc('eth_getTransactionCount', async (params: any) => {
  */
 app.useRpc('eth_call', async (params: any) => {
   return logAndHandleResponse('eth_call', params, (requestId) =>
-    relay.eth().call(params?.[0], params?.[1], requestId));
+      relay.eth().call(params?.[0], params?.[1], requestId));
 });
 
 /**
@@ -365,7 +365,7 @@ app.useRpc('eth_call', async (params: any) => {
  */
 app.useRpc('eth_sendRawTransaction', async (params: any) => {
   return logAndHandleResponse('eth_sendRawTransaction', params, (requestId) =>
-  relay.eth().sendRawTransaction(params?.[0], requestId));
+      relay.eth().sendRawTransaction(params?.[0], requestId));
 });
 
 /**
@@ -376,7 +376,7 @@ app.useRpc('eth_sendRawTransaction', async (params: any) => {
  */
 app.useRpc('eth_getTransactionReceipt', async (params: any) => {
   return logAndHandleResponse('eth_getTransactionReceipt', params, (requestId) =>
-    relay.eth().getTransactionReceipt(params?.[0], requestId));
+      relay.eth().getTransactionReceipt(params?.[0], requestId));
 });
 
 app.useRpc('web3_clientVersion', async () => {
@@ -400,7 +400,7 @@ app.useRpc('eth_accounts', async () => {
  */
 app.useRpc('eth_getTransactionByHash', async (params: any) => {
   return logAndHandleResponse('eth_getTransactionByHash', params, (requestId) =>
-    relay.eth().getTransactionByHash(params[0], requestId));
+      relay.eth().getTransactionByHash(params[0], requestId));
 });
 
 /**
@@ -417,7 +417,7 @@ app.useRpc('eth_getTransactionByHash', async (params: any) => {
  */
 app.useRpc('eth_feeHistory', async (params: any) => {
   return logAndHandleResponse('eth_feeHistory', params, (requestId) =>
-    relay.eth().feeHistory(Number(params?.[0]), params?.[1], params?.[2], requestId));
+      relay.eth().feeHistory(Number(params?.[0]), params?.[1], params?.[2], requestId));
 });
 
 /**
@@ -428,7 +428,7 @@ app.useRpc('eth_feeHistory', async (params: any) => {
  */
 app.useRpc('eth_getBlockTransactionCountByHash', async (params: any) => {
   return logAndHandleResponse('eth_getBlockTransactionCountByHash', params, (requestId) =>
-    relay.eth().getBlockTransactionCountByHash(params?.[0], requestId));
+      relay.eth().getBlockTransactionCountByHash(params?.[0], requestId));
 });
 
 /**
@@ -439,7 +439,7 @@ app.useRpc('eth_getBlockTransactionCountByHash', async (params: any) => {
  */
 app.useRpc('eth_getBlockTransactionCountByNumber', async (params: any) => {
   return logAndHandleResponse('eth_getBlockTransactionCountByNumber', params, (requestId) =>
-    relay.eth().getBlockTransactionCountByNumber(params?.[0], requestId));
+      relay.eth().getBlockTransactionCountByNumber(params?.[0], requestId));
 });
 
 /**
@@ -452,12 +452,12 @@ app.useRpc('eth_getLogs', async (params: any) => {
   const filter = params[0];
 
   return logAndHandleResponse('eth_getLogs', params, (requestId) => relay.eth().getLogs(
-    filter.blockHash,
-    filter.fromBlock,
-    filter.toBlock,
-    filter.address,
-    filter.topics,
-    requestId
+      filter.blockHash,
+      filter.fromBlock,
+      filter.toBlock,
+      filter.address,
+      filter.topics,
+      requestId
   ));
 });
 
@@ -472,7 +472,7 @@ app.useRpc('eth_getLogs', async (params: any) => {
  */
 app.useRpc('eth_getStorageAt', async (params: any) => {
   return logAndHandleResponse('eth_getStorageAt', params, (requestId) =>
-    relay.eth().getStorageAt(params?.[0], params?.[1], params?.[2], requestId));
+      relay.eth().getStorageAt(params?.[0], params?.[1], params?.[2], requestId));
 });
 
 /**
@@ -484,7 +484,7 @@ app.useRpc('eth_getStorageAt', async (params: any) => {
  */
 app.useRpc('eth_getTransactionByBlockHashAndIndex', async (params: any) => {
   return logAndHandleResponse('eth_getTransactionByBlockHashAndIndex', params, (requestId) =>
-    relay.eth().getTransactionByBlockHashAndIndex(params?.[0], params?.[1], requestId));
+      relay.eth().getTransactionByBlockHashAndIndex(params?.[0], params?.[1], requestId));
 });
 
 /**
@@ -496,7 +496,7 @@ app.useRpc('eth_getTransactionByBlockHashAndIndex', async (params: any) => {
  */
 app.useRpc('eth_getTransactionByBlockNumberAndIndex', async (params: any) => {
   return logAndHandleResponse('eth_getTransactionByBlockNumberAndIndex', params, (requestId) =>
-    relay.eth().getTransactionByBlockNumberAndIndex(params?.[0], params?.[1], requestId));
+      relay.eth().getTransactionByBlockNumberAndIndex(params?.[0], params?.[1], requestId));
 });
 
 /**
@@ -510,7 +510,7 @@ app.useRpc('eth_getTransactionByBlockNumberAndIndex', async (params: any) => {
  */
 app.useRpc('eth_getUncleByBlockHashAndIndex', async () => {
   return logAndHandleResponse('eth_getUncleByBlockHashAndIndex', [], (requestId) =>
-    relay.eth().getUncleByBlockHashAndIndex(requestId));
+      relay.eth().getUncleByBlockHashAndIndex(requestId));
 });
 
 /**
@@ -523,7 +523,7 @@ app.useRpc('eth_getUncleByBlockHashAndIndex', async () => {
  */
 app.useRpc('eth_getUncleByBlockNumberAndIndex', async () => {
   return logAndHandleResponse('eth_getUncleByBlockNumberAndIndex', [], (requestId) =>
-    relay.eth().getUncleByBlockNumberAndIndex(requestId));
+      relay.eth().getUncleByBlockNumberAndIndex(requestId));
 });
 
 /**
@@ -535,7 +535,7 @@ app.useRpc('eth_getUncleByBlockNumberAndIndex', async () => {
  */
 app.useRpc('eth_getUncleCountByBlockHash', async () => {
   return logAndHandleResponse('eth_getUncleCountByBlockHash', [], (requestId) =>
-    relay.eth().getUncleCountByBlockHash(requestId));
+      relay.eth().getUncleCountByBlockHash(requestId));
 });
 
 /**
@@ -547,7 +547,7 @@ app.useRpc('eth_getUncleCountByBlockHash', async () => {
  */
 app.useRpc('eth_getUncleCountByBlockNumber', async () => {
   return logAndHandleResponse('eth_getUncleCountByBlockNumber', [], (requestId) =>
-    relay.eth().getUncleCountByBlockNumber(requestId));
+      relay.eth().getUncleCountByBlockNumber(requestId));
 });
 
 /**
@@ -630,12 +630,12 @@ app.useRpc('eth_maxPriorityFeePerGas', async () => {
 
 app.useRpc('eth_newFilter', async (params: any) => {
   const filter = params[0];
-  return logAndHandleResponse('eth_newFilter', [], (requestId) => relay.eth().newFilter(
-    filter?.fromBlock,
-    filter?.toBlock,
-    filter?.address,
-    filter?.topics,
-    requestId));
+  return logAndHandleResponse('eth_newFilter', [], (requestId) => relay.eth().filterService().newFilter(
+      filter?.fromBlock,
+      filter?.toBlock,
+      filter?.address,
+      filter?.topics,
+      requestId));
 });
 
 /**
@@ -673,7 +673,7 @@ app.useRpc('eth_coinbase', async () => {
  */
 app.useRpc('eth_uninstallFilter', async (params: any) => {
   return logAndHandleResponse('eth_uninstallFilter', params, (requestId) =>
-    relay.eth().filterService().uninstallFilter(params?.[0], requestId));
+      relay.eth().filterService().uninstallFilter(params?.[0], requestId));
 });
 
 const rpcApp = app.rpcApp();
